+++ webrtc/modules/audio_coding/codecs/isac/fix/interface/audio_encoder_isacfix.h
@@ -25,12 +25,12 @@ struct IsacFix {
   static const uint16_t kFixSampleRate = 16000;
   static inline int16_t Control(instance_type* inst,
                                 int32_t rate,
-                                int framesize) {
+                                int16_t framesize) {
     return WebRtcIsacfix_Control(inst, rate, framesize);
   }
   static inline int16_t ControlBwe(instance_type* inst,
                                    int32_t rate_bps,
-                                   int frame_size_ms,
+                                   int16_t frame_size_ms,
                                    int16_t enforce_frame_size) {
     return WebRtcIsacfix_ControlBwe(inst, rate_bps, frame_size_ms,
                                     enforce_frame_size);
@@ -38,11 +38,11 @@ struct IsacFix {
   static inline int16_t Create(instance_type** inst) {
     return WebRtcIsacfix_Create(inst);
   }
-  static inline int DecodeInternal(instance_type* inst,
-                                   const uint8_t* encoded,
-                                   int16_t len,
-                                   int16_t* decoded,
-                                   int16_t* speech_type) {
+  static inline int16_t DecodeInternal(instance_type* inst,
+                                       const uint8_t* encoded,
+                                       int16_t len,
+                                       int16_t* decoded,
+                                       int16_t* speech_type) {
     return WebRtcIsacfix_Decode(inst, encoded, len, decoded, speech_type);
   }
   static inline int16_t DecodePlc(instance_type* inst,
@@ -53,9 +53,9 @@ struct IsacFix {
   static inline int16_t DecoderInit(instance_type* inst) {
     return WebRtcIsacfix_DecoderInit(inst);
   }
-  static inline int Encode(instance_type* inst,
-                           const int16_t* speech_in,
-                           uint8_t* encoded) {
+  static inline int16_t Encode(instance_type* inst,
+                               const int16_t* speech_in,
+                               uint8_t* encoded) {
     return WebRtcIsacfix_Encode(inst, speech_in, encoded);
   }
   static inline int16_t EncoderInit(instance_type* inst, int16_t coding_mode) {
