+++ webrtc/modules/audio_coding/codecs/isac/main/source/isac_unittest.cc
@@ -80,7 +80,7 @@ TEST_F(IsacTest, DISABLED_ON_IOS(IsacUpdateBWE)) {
   WebRtcIsac_EncoderInit(isac_codec_, 0);
   WebRtcIsac_DecoderInit(isac_codec_);
 
-  int encoded_bytes;
+  int16_t encoded_bytes;
 
   // Test with call with a small packet (sync packet).
   EXPECT_EQ(-1, WebRtcIsac_UpdateBwEstimate(isac_codec_, bitstream_small_, 7, 1,
