+++ webrtc/modules/audio_coding/codecs/isac/main/test/SwitchingSampRate/SwitchingSampRate.cc
@@ -191,7 +191,7 @@ int main(int argc, char* argv[])
 
   short streamLen;
   short numSamplesRead;
-  int lenDecodedAudio;
+  short lenDecodedAudio;
   short senderIdx;
   short receiverIdx;
